# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/trxfee/list": {
            "get": {
                "description": "get trx fee by given time period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of trx fee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "start timestamp",
                        "name": "start_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "end timestamp",
                        "name": "end_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page starting from 0",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "20 by default",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GetTrxFeeListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trxfee/{trx_hash}": {
            "get": {
                "description": "get trx fee by trx hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get trx fee of single trx",
                "parameters": [
                    {
                        "type": "string",
                        "description": "trx hash",
                        "name": "trx_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "repository.UniTrxFee": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "integer"
                },
                "ethUsdtPrice": {
                    "type": "number"
                },
                "gasPrice": {
                    "type": "integer"
                },
                "gasUsed": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "trxFeeUsdt": {
                    "type": "number"
                },
                "trxHash": {
                    "type": "string"
                },
                "trxTime": {
                    "type": "integer"
                }
            }
        },
        "service.GetTrxFeeListResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.UniTrxFee"
                    }
                }
            }
        }
    }
}